refactor(core): move sourceMapValue to separate file
